Attempting to fix... hangs?
I started some work on megassh but can't remember exactly what I was trying to do.  I think I was seeing some commands that produce a lot of output hang.  I usually don't have problems with the command but they're not generating tremendous amounts of data.

I'm going to stash the changes in a new branch rather than commit to master for now.