Improving parsing of target
My script was not allowing for a filename to contain a colon and I wanted to improve that.  I refactored how the target is parsed so you had to have a colon separate the remote system and the remote filename but it doesn't matter if there are colons in the filename'.